+++ src/App.js
@@ -3,7 +3,9 @@ import './App.css';
 import ComfyJS from 'comfy.js'
 import { GiphyFetch } from '@giphy/js-fetch-api'
 
-import { useMemo, useState, useEffect, useRef } from 'react'
+import { useMemo, useState, useEffect } from 'react'
+
+import { MediaPlayer } from './MediaPlayer';
 
 const logger = console;
 
@@ -109,48 +111,6 @@ function TwitchContainer({ giphyKey, twitchChannel }) {
 
   return (
     <MediaPlayer url={url} />
-  );
-}
-
-function MediaPlayer({ url }) {
-  const videoRef = useRef(null);
-
-  useEffect(() => {
-    if(!videoRef.current) return;
-    if(!url) return;
-
-    (async function() {
-      videoRef.current.src = url;
-      await videoRef.current.load();
-      videoRef.current.play();
-
-      let onend = () => {};
-
-      let MINIMUM_DURATION = 5 * 1000;
-
-      let startTime = new Date().getTime();
-
-      onend = (e) => {
-        let endTime = new Date().getTime();
-
-        let duration = endTime - startTime;
-
-        if(duration <= MINIMUM_DURATION) {
-          logger.info("Duration not reached looping");
-          videoRef.current.play();
-        } else {
-          logger.info("Duration reached ending");
-          videoRef.current.src = null;
-          videoRef.current.removeEventListener("ended", onend);
-        }
-      };
-
-      videoRef.current.addEventListener("ended", onend);
-    })();
-  }, [ videoRef, url ]);
-
-  return (
-    <video ref={videoRef}></video>
   );
 }
